add a copy of lxtask (LXDE Task Manager)
This is a copy of lxtask I built myself.

I cloned the source code from https://github.com/lxde/lxtask and built it
with the following commands at commit lxtask (master)>g log .
commit 3872237d397a ("Support silent building"):

> ./autogen.sh
> ./configure
> make

After building it, I installed it running:

> sudo make install

And it was installed at /usr/local/bin/lxtask.

The version built was 0.1.10.

lxtask is derived from xfce4-taskmanager, but with dependencies on
xfce removed, and improvements made.

I'm adding it, because I want a replacement on Linux for the
Windows Task Manager.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>