add a copy of ksuperkey
ksuperkey is a workaround a limitation in Xfce. It allows us to open
the Start menu using the Super key (also known as "Meta" or
"the Windows key"). If we hold down the Super key it will still
act as a modifier key, allowing us to use it for other keyboard
shortcuts. As without this workaround, our shortcut like
Super + R or Super + E wouldn't (they do not) work.

Building ksuperkey is very simple, just clone and build it like:

> git clone https://github.com/hanschen/ksuperkey.git
> cd ksuperkey
> make

For installing it system-wide, run:

> sudo make install

On my system I have a Ctrl+Esc keyboard shortcut
to open the Start menu.

To test ksuperkey, run the following:

> ksuperkey -e "Super_L=Control_L|Escape"

Now pressing and releasing Super will open the Start menu on Xfce. Bingo!

Signed-off-by: Thiago Farina <tfransosi@gmail.com>