add a script to take a screenshot of login screen
This script only works for lightdm display manager.

The next step is to switch to one of the virtual terminals (tty).
Press Ctrl+Alt+1 (up to 6) to switch.

Enter your username and password to login, then enter this
command:

$ cd ~/bin
$ sudo ./loginscreen

Stay in the login screen for 5 seconds at least as the time
delay for capturing the screen was set to this time.
After that, go back to the login screen pressing Ctrl+Alt+7,
and login again in your desktop. Go to the home directory,
and if everything worked properly there will be a file named
login-screenshot.png in there. If that is the case, open it
with your image viewer to see the screenshot of your login screen.

All my thanks goes to https://vitux.com/how-to-take-screenshots-of-lock-and-login-screens-in-ubuntu/
for the script code.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>